+++ src/node.cpp
@@ -5,7 +5,7 @@ namespace bezier
     Node Node::operator+(const Node& other) const {
         return {x + other.x, y + other.y};
     }
-    Node Node::operator*(double scalar) const {
+    Node Node::operator*(float scalar) const {
         return {x * scalar, y * scalar};
     }
